server/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
tensorflow-cpu
faiss-cpu
langchain
langchain-community
langchain-mistralai
sentence-transformers
mistralai
huggingface-hub
pydantic
python-multipart
python-dotenv
